Updated
OK this is the last version.  I realized I forgot to do a merge at the end.  I created 2 variables and did merges for both: The first is the one I told you about- the construction permits to house units ratio ad  the second is just number of permits in each tract.  I thought the second one might be useful in case we wanted to see if any of the census variables correlated with the number of permits.  I run that correlation as my last command and you can see that construction permits are most highly correlated with FAGI total 2013.  The correlation between permits and number of housing units was .20 (H0010002 = total housing units)